% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordex_grid_cell_value_from_points.R
\name{cordex_grid_cell_value_from_points}
\alias{cordex_grid_cell_value_from_points}
\title{Extraction of CORDEX grid  values from points}
\usage{
cordex_grid_cell_value_from_points(
  nc,
  crop_sf,
  point_id_name = c("STATION_NAMES"),
  names_xy = c("x", "y"),
  type = c("grid", "point"),
  ...
)
}
\arguments{
\item{nc}{file (or connection to)  provided by CORDEX / EURO-CORDEX projections}

\item{crop_sf}{spatial feature containing the points and on which the netcdf data has been cropped}

\item{point_id_name}{point id names in \code{crop_sf}.}

\item{names_xy}{names for x and y columns for the function value. Default is \code{c("x","y")}.}

\item{type}{option. Default is \code{c("grid","point")[1]} (only first element, \code{"grid"} is considered). See function internal code for details.}

\item{...}{further arguments to be implemented}
}
\description{
Extraction of CORDEX grid  values from points
}
\details{
CORDEX data are available on \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/projections-cordex-domains-single-levels?tab=overview}
}
\examples{

nc <- system.file('netcdfs/pr_EUR-11_CCCma-CanESM2_rcp85_r1i1p1_CLMcom-CCLM4-8-17_v1_mon_208101-209012.nc',package="cordexr")

library(RMAWGEN)
data(trentino)
stations <- cbind(as.data.frame(STATION_LATLON),STATION_NAMES) \%>\% st_as_sf(coords=c(1,2),crs=4326)

cc <- system.time(out <- cordex_grid_cell_value_from_points(nc,crop_sf=stations))

ccp <- system.time(outp <- cordex_grid_cell_value_from_points(nc,crop_sf=stations,type="point"))





}
\seealso{
\code{\link{cordex_grid_cell_values}},\code{\link{cordex_grid_vertices}}
}
