% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordex_grid_cell_values.R
\name{cordex_grid_cell_values}
\alias{cordex_grid_cell_values}
\title{Extraction of CORDEX grid  values}
\usage{
cordex_grid_cell_values(
  nc,
  rlon,
  rlat,
  x = rlon,
  y = rlat,
  time = NULL,
  variable = NULL,
  tz = "GMT",
  dim_x = c("rlon", "x", "auto"),
  dim_y = c("rlat", "x", "auto"),
  returns_time_posixct = TRUE,
  add_suffix = as.character(NA),
  ncatts_in_df = FALSE,
  latlon_tolerance = 0.01,
  lonlat_tolerance = latlon_tolerance,
  lon_tolerance = lonlat_tolerance,
  lat_tolerance = lonlat_tolerance,
  time_tolerance = 10/(24 * 3600),
  ...
)
}
\arguments{
\item{nc}{file (or connection to)  provided by CORDEX / EURO-CORDEX projections}

\item{rlon, x}{rotated longitude values or x axis values}

\item{rlat, y}{rotated latitude values or y axis values}

\item{time}{time coordinate  values}

\item{variable}{variable, e.g. c("pr","tas","tasmin","tasmax") or \code{NULL}. In case of \code{NULL} (default) , it is retrived of \code{nc} filename. See \code{varid} on \code{\link{ncvar_get}}}

\item{tz}{time zone used for time unit. Not specified in the CORDEX netcdf file. Default is \code{GMT}}

\item{dim_x}{possible names for x/longitude dimension}

\item{dim_y}{possible names for y/latitude dimension}

\item{returns_time_posixct}{logical value. If it is \code{TRUE} (default) , time coordinate will be transformed and returned as POSIxct class (\code{\link{as.POSIXct}}).}

\item{add_suffix}{suffix that can be  added to the the name variable. Default is \code{NA} (no suffix added).}

\item{ncatts_in_df}{logical value. If it is \code{TRUE} netCDF attributes will be added to the returened data frame.}

\item{latlon_tolerance, lonlat_tolerance, lon_tolerance, lat_tolerance, time_tolerance}{tolarances}

\item{...}{further arguments.}
}
\description{
Extraction of CORDEX grid  values
}
\details{
CORDEX data are available on \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/projections-cordex-domains-single-levels?tab=overview}
}
\examples{

nc <- system.file('netcdfs/pr_EUR-11_CCCma-CanESM2_rcp85_r1i1p1_CLMcom-CCLM4-8-17_v1_mon_208101-209012.nc',package="cordexr")

library(RMAWGEN)
data(trentino)
stations <- cbind(as.data.frame(STATION_LATLON),STATION_NAMES) \%>\% st_as_sf(coords=c(1,2),crs=4326)
rlatlon <- cordex_grid_vertices(nc,crop_sf=stations)

out <- cordex_grid_cell_values(nc,rlon=rlatlon$rlon,rlat=rlatlon$rlat)
out2 <- cordex_grid_cell_values(nc,rlon=rlatlon$rlon,rlat=rlatlon$rlat,returns_time_posixct=FALSE)



}
\seealso{
\code{\link{cordex_grid_vertices}},\code{\link{ncvar_get}}
}
