% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex_grid_points.R
\name{cordex_grid_points}
\alias{cordex_grid_points}
\title{Extraction of CORDEX grid centers of cells (lat/lon coordinates)}
\usage{
cordex_grid_points(
  nc,
  ...,
  dim_x = c("rlon", "x", "auto"),
  dim_y = c("rlat", "y", "auto"),
  crs = 4326,
  return_sf = TRUE,
  crop_sf = NULL
)
}
\arguments{
\item{nc}{file (or connection to)  provided by CORDEX / EURO-CORDEX projections}

\item{...}{further arguments}

\item{dim_x}{possible names for x/longitude dimension}

\item{dim_y}{possible names for y/latitude dimension}

\item{crs}{CRS for output. Used in case \code{return_sf==TRUE} or \code{crop_sf} is not null.}

\item{return_sf}{logical variable. If it is \code{TRUE} , it return \code{\link{sf}} object.}

\item{crop_sf}{spatial feature or an extent on which the netcdf data has been cropped.}
}
\description{
Extraction of CORDEX grid centers of cells (lat/lon coordinates)
}
\examples{

nc <- system.file('netcdfs/pr_EUR-11_CCCma-CanESM2_rcp85_r1i1p1_CLMcom-CCLM4-8-17_v1_mon_208101-209012.nc',package="cordexr")
out <- cordex_grid_points(nc)

library(RMAWGEN)
data(trentino)
stations <- cbind(as.data.frame(STATION_LATLON),STATION_NAMES) \%>\% st_as_sf(coords=c(1,2),crs=4326)
out <- cordex_grid_points(nc,crop_sf=stations)



}
\seealso{
\url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/projections-cordex-domains-single-levels?tab=overview})
}
