% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex_grid_vertices.R
\name{cordex_grid_vertices}
\alias{cordex_grid_vertices}
\title{Extraction of CORDEX grid  cells (lat/lon coordinates of the polygons)}
\usage{
cordex_grid_vertices(
  nc,
  ...,
  dim_x = c("rlon", "x", "auto"),
  dim_y = c("rlat", "y", "auto"),
  lon_vertices = c("lon_vertices", "bounds_lon", "rlon_bnds", "vertices_longitude",
    "auto"),
  lat_vertices = c("lat_vertices", "bounds_lat", "vertices_latitude", "auto"),
  nvertex = 4,
  crs = 4326,
  return_sf = TRUE,
  crop_sf = NULL,
  buffer = 0.1
)
}
\arguments{
\item{nc}{file (or connection to)  provided by CORDEX / EURO-CORDEX projections}

\item{...}{further arguments}

\item{dim_x}{possible names for x/longitude dimension}

\item{dim_y}{possible names for y/latitude dimension}

\item{lon_vertices, lat_vertices}{possible names for possible names of grid vertex longitude and latitude.}

\item{nvertex}{number of grid vertices. Default is 4.}

\item{crs}{CRS for output. Used in case \code{return_sf==TRUE} or \code{crop_sf} is not null.}

\item{return_sf}{logical variable. If it is \code{TRUE} , it return \code{\link{sf}} object.}

\item{crop_sf}{spatial feature or an extent on which the netcdf data has been cropped.}

\item{buffer}{buffer in coordinate unit (e.g. lat/lon degrees) in order to increase the extent or bounding box of \code{crop_sf}.}
}
\description{
Extraction of CORDEX grid  cells (lat/lon coordinates of the polygons)
}
\examples{

nc <- system.file('netcdfs/pr_EUR-11_CCCma-CanESM2_rcp85_r1i1p1_CLMcom-CCLM4-8-17_v1_mon_208101-209012.nc',package="cordexr")
out <- cordex_grid_vertices(nc)


\dontrun{
## mapview::mapview(out)
}

library(RMAWGEN)
data(trentino)
stations <- cbind(as.data.frame(STATION_LATLON),STATION_NAMES) \%>\% st_as_sf(coords=c(1,2),crs=4326)
out <- cordex_grid_vertices(nc,crop_sf=stations)

plot(st_geometry(out))
plot(stations,add=TRUE)
## LOOK AT ST_WITHIN
station_cell <- st_within(stations,out,sparse=TRUE) \%>\% sapply(function(x){x[1]}) ## see buffer
stations$nccell <- station_cell

out2 <- as.data.frame(out[station_cell,]) \%>\% cbind(stations)

station_cell <- st_within(stations,out) \%>\% unlist()## see buffer
\dontrun{
ncs <- '/home/ecor/local/data/climate/cordex/europe/monthly_unzipped/' \%>\% list.files(full.name=TRUE,pattern=".nc")
##ncs[-c(22,24,25,37,40,47,49,52,56,60,62,65,83,86,89,92,104,107,114,116,119,123,127,129,132,154,156,157,169,172,179,181,184,188,192,194,197,219,222,234,237,244,246,249,253,257,259,262)

}
}
