% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordex_netcdf_attributes.R
\name{cordex_netcdf_attributes}
\alias{cordex_netcdf_attributes}
\title{Extraction of CORDEX netcdf attributes}
\usage{
cordex_netcdf_attributes(
  nc,
  variable = NULL,
  ncatts_in_df = FALSE,
  ncatts_in_df_names = c("driving_model_id", "model_id", "institution",
    "experiment_id", "variable", "frequency", "source", "driving_model_ensemble_member"),
  ...
)
}
\arguments{
\item{nc}{file (or connection to)  provided by CORDEX / EURO-CORDEX projections}

\item{variable}{variable, e.g. c("pr","tas","tasmin","tasmax") or \code{NULL}. In case of \code{NULL} (default) , it is retrived of \code{nc} filename. See \code{varid} on \code{\link{ncvar_get}}}

\item{ncatts_in_df}{logical value. If it is \code{TRUE} netCDF attributes will be added to the returened data frame.}

\item{ncatts_in_df_names}{names}

\item{...}{further arguments.}
}
\description{
Extraction of CORDEX netcdf attributes
}
\details{
CORDEX data are available on \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/projections-cordex-domains-single-levels?tab=overview}
}
\examples{

nc <- system.file('netcdfs/pr_EUR-11_CCCma-CanESM2_rcp85_r1i1p1_CLMcom-CCLM4-8-17_v1_mon_208101-209012.nc',package="cordexr")
out <- cordex_netcdf_attributes(nc)
out <- cordex_netcdf_attributes(nc,ncatts_in_df=TRUE)


}
\seealso{
\code{\link{cordex_netcdf_attributes_multi_sources}},\code{\link{ncvar_get}}
}
