% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordex_netcdf_attributes_multi_sources.R
\name{cordex_netcdf_attributes_multi_sources}
\alias{cordex_netcdf_attributes_multi_sources}
\title{Extraction of CORDEX netcdf attributes from several Netcdf file sources}
\usage{
cordex_netcdf_attributes_multi_sources(nc, ...)
}
\arguments{
\item{nc}{file template name  provided by CORDEX / EURO-CORDEX projections. See examples}

\item{...}{further arguments for \code{\link{cordex_netcdf_attributes}}}
}
\description{
Extraction of CORDEX netcdf attributes from several Netcdf file sources
}
\examples{

ncs <- '/home/ecor/local/data/climate/cordex/europe/monthly_unzipped/' \%>\% list.files(full.name=TRUE,pattern=".nc")
out <- cordex_netcdf_attributes_multi_sources(ncs)



}
