% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_ncvar_name.R
\name{which_ncvar_name}
\alias{which_ncvar_name}
\title{Finds the real name of a variable in a netCDF file.}
\usage{
which_ncvar_name(nc, ncvar_name, ...)
}
\arguments{
\item{nc}{file (or connection to)  provided by CORDEX / EURO-CORDEX projections}

\item{ncvar_name}{possible names for the netCDF variables}

\item{...}{further arguments (not used)}
}
\description{
Finds the real name of a variable in a netCDF file.
}
\examples{

nc1 <- "/home/ecor/local/data/climate/cordex/europe/monthly_unzipped//pr_EUR-11_MOHC-HadGEM2-ES_historical_r1i1p1_ICTP-RegCM4-6_v1_mon_198101-199012.nc"
var <- which_ncvar_name(nc1,ncvar_name=c("X","x","rlon","auto"))


}
